requirements.txt
aiogram==3.4.1
g4f==0.2.8.2
moviepy==1.0.3
numpy==1.26.4
python-dotenv==1.0.1
Pillow==10.3.0
